ai_tg_bot/requirements.txt
faiss-cpu
numpy
telethon
openai
python-dotenv
notion-client